Exercises 4.15-4.23 completed; part4d

+++ part4/bloglist-backend/controllers/bloglist.js
@@ -1,15 +1,35 @@
 const bloglistRouter = require('express').Router()
 const Blog = require('../models/bloglist')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
+
 
 
 bloglistRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  //console.log('blogs: ', blogs)
   response.json(blogs)
+})
+
+bloglistRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).end()
+  }
 })
 
 
 bloglistRouter.post('/', async (request, response) => {
-  console.log('get new blog request')
+  // const decodedToken = jwt.verify(request.token, config.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+
   const body = request.body
   if (!body.likes) {
     body.likes = 0
@@ -20,11 +40,23 @@ bloglistRouter.post('/', async (request, response) => {
 
   const blog = new Blog(body)
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
 
 bloglistRouter.delete('/:id', async (request, response) => {
+  // const decodedToken = jwt.verify(request.token, config.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
